+++ src/options.ts
@@ -678,6 +678,10 @@ function translateKeys() {
 }
 
 function configurePuckSettings() {
+  if (!__ENABLE_PUCK__) {
+    return;
+  }
+
   const showPuckOptions = Array.from(
     document.querySelectorAll<HTMLInputElement>(
       'input[type=radio][name=showPuck]'
@@ -785,7 +789,9 @@ function fillVals() {
   optform.popupStyle.value = config.popupStyle;
   optform.tabDisplay.value = config.tabDisplay;
   optform.toolbarIcon.value = config.toolbarIcon;
-  optform.showPuck.value = config.showPuck;
+  if (__ENABLE_PUCK__) {
+    optform.showPuck.value = config.showPuck;
+  }
 
   getConfiguredToggleKeyValue()
     .then((toggleCommand) => {